+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Releases
 - Core modules have been cythonized for additional performance improvements.
 - **Breaking** All custom exceptions are exported by the ``thriftrw.errors``
   module only. This includes ``ThriftProtocolError`` and ``EndOfInputError``.
+- ``UnknownExceptionError`` is now raised if an unrecognized exception is
+  encountered while parsing service method responses.
 
 
 0.4.2 (2015-10-13)